Reorganize test_common
