+++ rust/parquet/src/util/test_common/rand_gen.rs
@@ -15,11 +15,11 @@
 // specific language governing permissions and limitations
 // under the License.
 
+
 use rand::{
     distributions::{range::SampleRange, Distribution, Standard},
     thread_rng, Rng,
 };
-use std::{env, fs, io::Write, path::PathBuf, str::FromStr};
 
 use crate::data_type::*;
 use crate::util::memory::ByteBufferPtr;
@@ -143,48 +143,4 @@ where
     for _ in 0..n {
         result.push(rng.gen_range(low, high));
     }
-}
-
-/// Returns path to the test parquet file in 'data' directory
-pub fn get_test_path(file_name: &str) -> PathBuf {
-    let result = env::var("PARQUET_TEST_DATA");
-    if result.is_err() {
-        panic!("Please point PARQUET_TEST_DATA environment variable to the test data directory");
-    }
-    let mut pathbuf = PathBuf::from_str(result.unwrap().as_str()).unwrap();
-    pathbuf.push(file_name);
-    pathbuf
-}
-
-/// Returns file handle for a test parquet file from 'data' directory
-pub fn get_test_file(file_name: &str) -> fs::File {
-    let file = fs::File::open(get_test_path(file_name).as_path());
-    if file.is_err() {
-        panic!("Test file {} not found", file_name)
-    }
-    file.unwrap()
-}
-
-/// Returns file handle for a temp file in 'target' directory with a provided content
-pub fn get_temp_file(file_name: &str, content: &[u8]) -> fs::File {
-    // build tmp path to a file in "target/debug/testdata"
-    let mut path_buf = env::current_dir().unwrap();
-    path_buf.push("target");
-    path_buf.push("debug");
-    path_buf.push("testdata");
-    fs::create_dir_all(&path_buf).unwrap();
-    path_buf.push(file_name);
-
-    // write file content
-    let mut tmp_file = fs::File::create(path_buf.as_path()).unwrap();
-    tmp_file.write_all(content).unwrap();
-    tmp_file.sync_all().unwrap();
-
-    // return file handle for both read and write
-    let file = fs::OpenOptions::new()
-        .read(true)
-        .write(true)
-        .open(path_buf.as_path());
-    assert!(file.is_ok());
-    file.unwrap()
 }
